+++ NEWS.md
@@ -1,3 +1,6 @@
+- Added `as.nlist.numeric()` to convert flattened nlist objects back into original objects.
+- Added `unlist.nlist()` to flatten nlist objects into term named numeric vectors.
+
 # nlist 0.0.1
 
 - Initial release.